+++ packages/contracts/src/index.ts
@@ -3,6 +3,7 @@ export * from "./approval.js";
 export * from "./common.js";
 export * from "./desktop-ipc.js";
 export * from "./events.js";
+export * from "./knowledge.js";
 export * from "./model.js";
 export * from "./run.js";
 export * from "./runtime-rpc.js";